refactor(dot reporter): extract coloring functions
so it's easier to test'em and the dot main module is more readable
