+++ test/report/dot/dot.spec.js
@@ -1,11 +1,11 @@
 "use strict";
 const fs               = require('fs');
 const expect           = require('chai').expect;
-const render           = require('../../src/report/dot');
-const deps             = require('./fixtures/cjs-no-dependency-valid.json');
-const unresolvableDeps = require('./fixtures/es6-unresolvable-deps.json');
-const doNotFollowDeps  = require('./fixtures/do-not-follow-deps.json');
-const orphanDeps       = require('./fixtures/orphan-deps.json');
+const render           = require('../../../src/report/dot');
+const deps             = require('../fixtures/cjs-no-dependency-valid.json');
+const unresolvableDeps = require('../fixtures/es6-unresolvable-deps.json');
+const doNotFollowDeps  = require('../fixtures/do-not-follow-deps.json');
+const orphanDeps       = require('../fixtures/orphan-deps.json');
 
 const clusterlessFixture  = fs.readFileSync('test/report/fixtures/clusterless.dot', 'utf8');
 const unresolvableFixture = fs.readFileSync('test/report/fixtures/unresolvable.dot', 'utf8');
